+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.1.0
 
+* Removed debug log message (thanks evincarofautumn!)
 * Minor documentation fixes
 
 ## 0.1.0.0 -- 2023-01-07