+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Removed debug log message (thanks evincarofautumn!)
 * Minor documentation fixes
 * Added quotes to variables in error messages
+* Added test for `-XQualifiedDo`
 
 ## 0.1.0.0 -- 2023-01-07